requirements.txt
asgiref==3.5.2
boto3==1.26.14
botocore==1.29.14
jmespath==1.0.1
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
